% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitSteamReviews.R
\name{split_steam_reviews}
\alias{split_steam_reviews}
\title{Splitting review into overall rating, # of reviews, and \% of positive reviews}
\usage{
split_steam_reviews(data, time)
}
\arguments{
\item{data}{A tibble of data with a column whose name includes "_reviews".
The data should be formatted like in steam_games.
Named because it's a tibble of data.}

\item{time}{A string indicating review column to split. Ex: "all", "recent".
This string will be appended to the start of the "_rating" and
"\emph{reviews" names, and the end of the "percent_positive}" name.
Named because it indicates what time the data is from.}
}
\value{
A tibble with the same columns as param data, except the chosen
"_reviews" column has been split into "_ratings", "_reviews", and
"percent_positive".
}
\description{
This function splits a column of review data derived from the steam_games data
set into three pieces of data via delimiters. These pieces are:
\itemize{
\item "_rating": Overall rating of game. String. Ex: Positive, Negative, Mixed.
\item "_reviews": Number of reviews game received. Numeric. Ex: 183, 18, 67392.
\item "percent_positive_": \% of reviews that are positive. Numeric. Ex: 98, 12, 3.
}
}
\details{
It is expected that the review data is formatted with specific delimiters
present in the steam_games data.
}
